sync up

Former-commit-id: 94a4b23bae7bcbb36052e436a5167c336bb7343e

+++ connect.go
@@ -221,7 +221,7 @@ func (c *Connection) runClient() error {
 			connection, err := net.Dial("tcp", c.Server+":"+port)
 			if err != nil {
 				if c.Server == "cowyo.com" {
-					fmt.Println("\nThe public server is down. Please tweet the webmaster: @yakczar")
+					fmt.Println("\nCheck http://bit.ly/croc-relay to see if the public server is down or contact the webmaster: @yakczar")
 				} else {
 					fmt.Printf("\nCould not connect to relay %s\n", c.Server)
 				}
@@ -332,7 +332,9 @@ func (c *Connection) runClient() error {
 						if id == 0 {
 							fmt.Printf("\n\nReceiving (<-%s)..\n", sendersAddress)
 						}
-						c.receiveFile(id, connection)
+						if err := c.receiveFile(id, connection); err != nil {
+							log.Error(errors.Wrap(err, "Problem receiving the file: "))
+						}
 					}
 				}
 			}
@@ -426,8 +428,13 @@ func (c *Connection) receiveFile(id int, connection net.Conn) error {
 	fileSizeInt, _ := strconv.Atoi(fileDataString)
 	chunkSize := int64(fileSizeInt)
 	logger.Debugf("chunk size: %d", chunkSize)
+	if chunkSize == 0 {
+		logger.Debug(fileSizeBuffer)
+		return errors.New("chunk size is empty!")
+	}
 
 	os.Remove(c.File.Name + ".enc." + strconv.Itoa(id))
+	log.Debug("Making " + c.File.Name + ".enc." + strconv.Itoa(id))
 	newFile, err := os.Create(c.File.Name + ".enc." + strconv.Itoa(id))
 	if err != nil {
 		panic(err)
@@ -486,7 +493,10 @@ func (c *Connection) sendFile(id int, connection net.Conn) error {
 		return err
 	}
 	logger.Debugf("sending chunk size: %d", fi.Size())
-	connection.Write([]byte(fillString(strconv.FormatInt(int64(fi.Size()), 10), 10)))
+	_, err = connection.Write([]byte(fillString(strconv.FormatInt(int64(fi.Size()), 10), 10)))
+	if err != nil {
+		return errors.Wrap(err, "Problem sending chunk data: ")
+	}
 
 	// show the progress
 	if !c.Debug {
